src/dashboard/backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
jinja2==3.1.6
python-multipart==0.0.20
httpx==0.28.1
python-dotenv==1.0.1